+++ examples/webvh/src/main.rs
@@ -106,9 +106,3 @@ impl From<anyhow::Error> for AppError {
         Self::Other(error)
     }
 }
-
-impl From<credibil_identity::Error> for AppError {
-    fn from(error: credibil_identity::Error) -> Self {
-        Self::Other(error.into())
-    }
-}
