+++ examples/webvh/src/update.rs
@@ -4,14 +4,10 @@ use axum::extract::State;
 use axum::{Json, http::StatusCode};
 use axum_extra::{TypedHeader, headers::Host};
 use credibil_identity::core::Kind;
-use credibil_identity::webvh::UpdateBuilder;
-use credibil_identity::{
-    DocumentBuilder, MethodType, PublicKeyFormat, VerificationMethod, VerificationMethodBuilder,
-    VmKeyId,
-};
-use credibil_identity::{
-    KeyPurpose,
-    webvh::{UpdateResult, resolve_log},
+use credibil_identity::did::{
+    DocumentBuilder, KeyPurpose, MethodType, PublicKeyFormat, VerificationMethod,
+    VerificationMethodBuilder, VmKeyId,
+    webvh::{UpdateBuilder, UpdateResult, resolve_log},
 };
 use serde::{Deserialize, Serialize};
