+++ src/did/key/resolve.rs
@@ -1,19 +1,21 @@
 //! Resolution of a DID into a public key using the `did:jwk` method.
 
-use crate::{Error, Method, MethodType, PublicKeyFormat, Resource, Url, VerificationMethod};
+use anyhow::bail;
+
+use crate::did::{Method, MethodType, PublicKeyFormat, Resource, Url, VerificationMethod};
 
 /// Convert a `did:key` URL into a [`VerificationMethod`] object.
-/// 
+///
 /// # Errors
 /// If the URL is not a valid `did:key` URL, an error is returned.
-pub fn resolve(url: &Url) -> crate::Result<Resource> {
+pub fn resolve(url: &Url) -> anyhow::Result<Resource> {
     if url.method != Method::Key {
-        return Err(Error::InvalidDid(format!("DID is not a valid did:key: {url}")));
+        bail!("DID is not a valid did:key: {url}");
     }
     // For did:key, the fragment is the key already multibase encoded. There is
     // no need to use a builder.
     let Some(fragment) = &url.fragment else {
-        return Err(Error::InvalidDid("DID is not a valid did:key - there is no fragment".into()));
+        bail!("DID is not a valid did:key - there is no fragment");
     };
     let vm = VerificationMethod {
         context: None,
