+++ src/did/web/did-eddsa.json
