+++ src/did/web/resolve.rs
@@ -1,6 +1,6 @@
 //! Resolve a DID URL to a DID Document for the `did:web` method.
 
-use crate::{DidResolver, Document, Error, Url};
+use crate::{Identity, IdentityResolver, did::{Document, Url}};
 
 impl Url {
     /// Convert a `did:web` URL to an HTTP URL pointing to the location of the
@@ -34,9 +34,10 @@ impl Url {
 /// # Errors
 /// If the URL cannot be converted to an HTTP format or if the resolver fails an
 /// error is returned.
-pub async fn resolve(url: &Url, resolver: &impl DidResolver) -> crate::Result<Document> {
+pub async fn resolve(url: &Url, resolver: &impl IdentityResolver) -> anyhow::Result<Document> {
     let http_url = url.to_web_http();
-    resolver.resolve(&http_url).await.map_err(|e| {
-        Error::InvalidDid(format!("issue resolving did:web: {e}"))
-    })
+    let id = resolver.resolve(&http_url).await?;
+    match id {
+        Identity::DidDocument(doc) => Ok(doc),
+    }
 }
